+++ src/visiomode_analysis/__init__.py
@@ -34,5 +34,6 @@ def cli():
 
 cli.add_command(session.session_cmd)
 cli.add_command(session.regressors_cmd)
+cli.add_command(session.session_start_time_cmd)
 cli.add_command(subject.subject_cmd)
 cli.add_command(group.group_cmd)